+++ src/cli.rs
@@ -5,6 +5,7 @@ use {
         terminal::{EnterAlternateScreen, LeaveAlternateScreen},
         QueueableCommand,
     },
+    directories_next::ProjectDirs,
     std::{
         env,
         fs,
@@ -41,6 +42,28 @@ pub fn run() -> Result<()> {
     let location = MissionLocation::new(&args)?;
     debug!("cargo_toml_file: {:?}", &location.cargo_toml_file);
 
+    let mut display_settings = DisplaySettings::default();
+
+    if let Some(project_dir) = ProjectDirs::from("org", "dystroy", "bacon") {
+        let prefs_path = project_dir.config_dir().join("prefs.toml");
+        if args.prefs {
+            if prefs_path.exists() {
+                println!("bacon prefs are at\n{:?}", &prefs_path);
+            } else {
+                fs::create_dir_all(prefs_path.parent().unwrap())?;
+                fs::write(&prefs_path, DEFAULT_PREFS)?;
+                println!(
+                    "{:?} file written.\nYou can modify it.",
+                    &prefs_path,
+                );
+            }
+            return Ok(());
+        }
+        let prefs = Prefs::from_path(&prefs_path)?;
+        info!("prefs: {:#?}", &prefs);
+        display_settings.apply_prefs(&prefs);
+    }
+
     let package_config_path = location.package_config_path();
     if args.init {
         if package_config_path.exists() {
@@ -51,20 +74,21 @@ pub fn run() -> Result<()> {
         } else {
             fs::write(&package_config_path, DEFAULT_PACKAGE_CONFIG)?;
             println!(
-                "{:?} written.\nYou can modify it.",
+                "{:?} file written.\nYou can modify it.",
                 &package_config_path,
             );
         }
         return Ok(())
     }
-
     let package_config = if package_config_path.exists() {
         PackageConfig::from_path(&package_config_path)?
     } else {
         PackageConfig::default()
     };
 
-    let display_settings = DisplaySettings::from(&args);
+    // args are applied after prefs, so that they can override them
+    display_settings.apply_args(&args);
+
     let mission = Mission::new(
         location,
         &package_config,