+++ src/default_prefs.rs
@@ -1,0 +1,17 @@
+
+pub static DEFAULT_PREFS: &str = r#"
+# This is a preferences file for the bacon tool
+# More info at https://github.com/Canop/bacon
+
+# Uncomment and change the value (true/false) to
+# specify whether bacon should start in summary mode
+#
+# summary = true
+
+# Uncomment and change the value (true/false) to
+# specify whether bacon should start with lines wrapped
+#
+# wrap = true
+
+"#;
+